fix(neon): fix unsafe blocks, clippy lints, and fuzz target (#3023)
## P0 CI Fix — Unblocks ALL PRs

Resolves 138+ clippy warnings that cause CI failures for every open PR.

### Changes (129 files)

**NEON kernel files (8 files):**

**Library code:**
- `bitnet-inference`: Collapse nested if, fix
field_reassign_with_default, remove useless vec!
- `bitnet-spirv`: Escape `OpenCL` in doc comment (doc_markdown lint)
- `bitnet-scoring-core`: Allow float_cmp in tests
- `simd_matmul.rs`: Gate `has_avx512f()` to `#[cfg(target_arch =
"x86_64")]` only

**Test files (110+ files):**
- Fix field_reassign_with_default, unnecessary clone/cast, approximate
PI
- Fix assert_eq with bool, borrowed expression, match for single pattern
- Add targeted `#[allow]` where appropriate

### Impact
- Unblocks 87+ Apple Silicon PRs and all other open PRs
- Zero functional changes — all fixes are lint compliance only
- Build and clippy verified clean locally

+++ tests/common/config_validator.rs
@@ -299,7 +299,8 @@ fn get_available_disk_space(_path: &Path) -> Option<u64> {
         let mut statvfs: libc::statvfs = unsafe { mem::zeroed() };
 
         if unsafe { libc::statvfs(path_cstr.as_ptr(), &mut statvfs) } == 0 {
-            Some(statvfs.f_bavail as u64 * statvfs.f_frsize as u64)
+            #[allow(clippy::unnecessary_cast)]
+            Some(statvfs.f_bavail as u64 * statvfs.f_frsize)
         } else {
             None
         }
